refactoring : 시간계산로직을 utils로 이동
